Update Table_39_Crude_Protein_Content_Swine_Feed_Provider.cs
Changed CP values to align with Denise's swine diet data. Please check that I didn't mess up the code.
Note: I have included the CP for the boar diet, but I am not sure if this has yet been added as a default diet in Holos.
Note: In the alg doc (Eq. 4.2.1-19), the CP is assumed to be a fraction (kg CP per kf feed), as the original data (from the Pork Protocol) were fractions; if the CP is currently reported as a percentage in the Holos code, let me know and I will add a divisor of 100 to this equation.