Update Table_33_Daily_Feed_Intake_For_Swine_Groups.csv
Aaron: I have added columns to specify the start and end weights for the different animal groups and the litter size (relevant to gilts and sows only). For the number of days in the management period, there is no entry for gilts and sows as there are multiple management periods for these groups which are not defined separately in the table (but the default number of days for each of these management periods is currently correct in the model interface for these groups).